+++ scrapy/extension.py
@@ -3,7 +3,7 @@ The Extension Manager
 
 See documentation in docs/topics/extensions.rst
 """
-from scrapy.middleware import MiddlewareManager
+from scrapy.middlewaremanager import MiddlewareManager
 from scrapy.utils.conf import build_component_list
 
 class ExtensionManager(MiddlewareManager):
